README.md
Horsechain Foundation Hub program. 

About the HorseChain Dev Program.

What is the HorseChain Hub program.?

HorseChain Foundation is aiming an extension of the offer program that would enable individuals to engage in longer term research and development projects supporting the HorseChain Foundation that using Cosmos SDK. The basic idea of the program is to provide Devlopers , with reliable funding on a monthly basis, thus enabling them to ‘go full-time’ on HorseChain.


What is the objective of Hub program?
    ●intended to encourage more longer-term and open-ended research projects. 
    ●Creating tools or providing solutions that offer scoped and immediate benefits and insights to the Transportation industry.
    ●Explore broader problems/Unsolved questions in the transportation . The program is suited to serving research needs. 
    ●Improve future decisions making. 

How does the HorseChain work with Dev. 
    ●Delivering value to the community. 
    ●Achieving Projects Milestones. 
    ●Having friendly open conversation with both entrepreneurs or groups. 

How will dev benefited?

    ●Each candidate will have agreement that makes them satisfied. 
    ●The agreement will be judged according to many development factors (such as customer experience, years and work experience). 


Applying to the HorseChain HUB. 

    ●There is no application to be submitted just get involved directly to our journey through our Github or reach us using Admin@horsechain.co